pyproject.toml
[tool.poetry]
name = "income-flow"
version = "0.1.0"
description = ""
authors = ["sieginglion <sieginglion@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dash = "^2.14.1"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
dash-bootstrap-components = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
